Move locators to the proper directory.
