+++ src/client/pythonEnvironments/base/locators/lowLevel/condaLocator.ts
