+++ src/client/pythonEnvironments/base/locators/lowLevel/windowsRegistryLocator.ts
