+++ src/client/pythonEnvironments/base/locators/lowLevel/windowsStoreLocator.ts
