+++ src/client/pythonEnvironments/common/commonUtils.ts
@@ -7,8 +7,8 @@ import { getOSType, OSType } from '../../common/utils/platform';
 import { PythonVersion } from '../base/info';
 import { parseExeVersion } from '../base/info/executable';
 import { compareVersions, getEmptyVersion, parseVersion } from '../base/info/pythonVersion';
-import { getPythonVersionFromConda } from '../discovery/locators/services/condaLocator';
-import { getPythonVersionFromVenv } from '../discovery/locators/services/virtualEnvironmentIdentifier';
+import { getPythonVersionFromConda } from '../base/locators/lowLevel/condaLocator';
+import { getPythonVersionFromVenv } from '../base/locators/lowLevel/virtualEnvironmentIdentifier';
 import { isDirectory, listDir } from './externalDependencies';
 import { isPosixPythonBin } from './posixUtils';
 import { isWindowsPythonExe } from './windowsUtils';
